Split steps into impl & definition to avoid browser bundling issues
Without this, the browser bundle has to include everything that's
referenced anywhere in the handler steps. Because the proxy step
creates RTC connections, that means it has to reference
node-datachannel and all sorts of other bits, which aren't relevant to
the browser bundle. Splitting into definitions and implementations
ensures that the browser only depends on the content necessary to define
and serialize each step, while the server has full access to everything.

This design in general matches the same structure in Mockttp, it's
pretty much a direction translation of the same solution to the same
problem.