Support connection renegotiation
This is supported by exposing a session API, returned on each connection
setup method (immediately with answerOffer or after setAnswer with
createOffer) that exposes the same methods but for mutating the existing
connection.

Seems like it might be interesting to have just one API in future:
peer.startSession(), then session.{everything} for initial setup &
renegotiation. Bit fiddly for now though, and slightly complicates the
non-renegotiated setup, so ignoring it for the moment.