Proxy media streams on hooked connections
Updates node-datachannel too, so that track.type() is available, and to
get fixes to avoid DC mid conflicts with tracks when creating offers.

This mirrors both the external & mock offer when they're created, so
that we include the required media streams throughout the connection.

Even with that, mirroring isn't sufficient: if the remote answer
doesn't include a data stream, then the resulting connection can't open
the control channel, so we don't successfully set up the proxy
connection.

To fix that, we now add an extra data stream if necessary for this case,
and we mirror correctly with that included to support this flow.