Add support for mirroring mock answer media params
This is a slightly more limited form of mirroring, since it's less
important, as libdatachannel mirrors most of the offer details for us
automatically.

The only key part is the SSRCs, which come from the answer side (i.e.
the offer & answer don't match) and which weren't included at all
beforehand, so proxied RTP packets weren't matched and were dropped.