Mirror our internal mock answer from the remote answer in hooks
This doesn't yet break any tests, but it's clearly correct: if we hook a
connection that sends an offer, and the real answer from the remote
client contains some media params, we should mirror those on the answer
from our mock peer, to ensure SSRCs etc are preserved throughout.

This also ensures that both connection completions happen in parallel,
rather than waiting unnecessarily for the two parts separately.