Fix flaky wait-for-track test, caused by race condition bug
Due to what seems to be a node/lib-datachannel bug, onTrack can fire
after conn.close() in some cases. This results in a segfault when we try
to use the resulting media track, so the workaround added here ignores
the event in the case that the connection is already shutdown.