Include user-agent metadata in peer-connected event
This could have been done via SDP, which would have been a slightly more
general solution perhaps, but that would require a lot more SDP munging,
and unmunging to avoid exposing it to remote proxied peers. Much better
to send it via the explicit API sidechannel, and this makes it easier to
use arbitrary data instead of just an attribute string.