Don't fire media stats events while no media is sending
This avoids traffic particularly in the case when the track is actually
dead, but nobody has realized yet (because of a silent shutdown - it is
UDP after all).

These are still fired every second, so you can assume any changes in a
given event happened within the preceeding second in any case.