Fix name of sdp conn event field, and expose explicit SessionDesc type
The field was named as SDP, but it's not just SDP - it's the entire
description, both SDP & type.

This introduces a new MockRTCSessionDescription, used for all output
APIs (but still accepting normal descriptions as input, to make the API
easily usable). This is better than using the built-in
RTCSessionDescriptionInit type since it makes 'sdp' non-optional,
limits the types to the ones relevant to us (no rollback/pranswer) and
explicitly exporting this makes it a little easier to use.