Add matching peer - initially proxying everything by default
This allows for a model more like Mockttp's (and like HTTP Toolkit's UI)
where you define a series of rules, and incoming connections run the
behaviour for the matching set of rules, instead of creating a different
peer to connection to for each possible behaviour.

No way to define custom rules yet - that'll come later, for now it's
just a peer that's always defined which will always proxy traffic
through.