Differentiate between track/channel created & opened events
Previously, the external 'openied' event was fired when the internal
stream was created, which was sometimes but not always the same as when
the track/chnnale actually opened. Instead, we now explicitly track the
two internally, so the external event may fire marginally later in
these cases, but will fire more accurately.