Use forceMediaTransport to fix issues with the latest NDC
More recent Libdatachannel versions include an optimization that
makes it impossible to add media tracks to an existing non-media
connection by default, unless 'forceMediaTransport' is enabled
to allow this. This option is only supported in the latest (0.4.3+)
node-datachannel, so here we update that and pass the option, fixing
the tests.

Also updates @types/node & TypeDoc very mildly, to avoid conflicts
with @types/ws transitive updates.